Update README.md
Hello! I have included my code below. This application is programmed to calculate the duration between two dates. As a result of calculating the years, months, and days the first few lines of coding provide a welcome statement to the calculator, instructional information, and gathers information for two dates. The following section of coding includes information for converting the string representation of the time and date  to the DateTime equivalent. Next, the coding for TimeSpan is utilized in order to take the difference between the two dates. Before concluding by running the program, the difference of the dates are broken into years, months, days, hours and minutes for formatting.